backend/services/judge_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
httpx==0.25.0
python-dotenv==1.0.0
pyjwt==2.8.1
sqlalchemy==2.0.23
asyncpg==0.29.0
requests==2.31.0
websockets==12.0

